Update worker.proposals.cpp
Refactor to revise voting thresholds for `eosio.saving` to match those of `eosio.amend`. Namely, change the thresholds to calculate as a percentage of "voteable TLOS" (VOTE token current supply) instead of number of unique accounts voting that contain any VOTE tokens. This reading is in compliance with the TBNOA and the previous version was out of compliance with that text.